requirements.txt
playwright==1.48.0
pyyaml==6.0.2
python-dotenv==1.0.1
schedule==1.2.2
msal==1.28.0
requests==2.32.3
beautifulsoup4==4.12.3

